+++ go.mod
@@ -1,3 +1,3 @@
 module http-server
 
-go 1.18
+go 1.24.0
